refactor(analytics): migrate QRScanner from useMetrics to useAnalytics
Deprecate MetaMetrics in favour of the shared analytics utility
(Batch 2-8 of MCWP-297).
